requirements.txt
langchain
langchain-ollama
langchain-openai
streamlit
langchain-core
langchain-community
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
fastapi
uvicorn
motor
bcrypt
pydantic
passlib
langgraph
python-jose
python-multipart
scipy
sounddevice
streamlit_extras
